fix problem with LL grammar and incorrect parsing of method calls.
The problem is that javacc is LL parser, but java language
is not LL language. For us, this shows up in the definition of
names versus method calls. For example
  java.lang.String.valueOf(...)
needs to be parsed as java.lang.String (the Name), and .valueOf (the
method, a Selector). Needed to add a LOOKAHEAD when parsing the
Name to make sure we don't take too many tokens. So, after pulling
off an ID, if the next tokens are DOT ID, but not DOT ID LPAREN,
then it's OK to pull of the DOT ID as the next part of the name.

Note that the example in the Javacc LOOKAHEAD MiniTutorial that
deals with this case (in the section "Semantic Lookahead") contains
an error. This was reported as a bug:
http://java.net/jira/browse/JAVACC-163
but was erroneously closed. I added a comment and a correction.